updated folder structure to separate out comms, fsm, and graphics
