--- src/main/java/swarm_wars_library/engine/EnVar.java
@@ -1,8 +1,0 @@
-package swarm_wars_library.engine;
-
-public class EnVar {
-
-    public final int screenWidth = 500;
-    public final int screenHeight = 500;
-
-}
