+++ src/main/java/swarm_wars_library/engine/Shooter.java
@@ -1,5 +1,6 @@
 /*Handles adding force to bullets, and setting bullet start location as Entity holding the shooter*/
 package swarm_wars_library.engine;
+
 import processing.core.PApplet;
 import java.util.*;
