+++ src/main/java/swarm_wars_library/engine/Tag.java
@@ -1,4 +1,5 @@
 package swarm_wars_library.engine;
+
 public enum Tag {
   EMPTY, PLAYER, ENEMY, E_BULLET, P_BULLET, ROCK, P_BOT, E_BOT; 
 }
