+++ src/main/java/swarm_wars_library/fsm/State.java
@@ -1,4 +1,5 @@
-package swarm_wars_library.engine;
+package swarm_wars_library.fsm;
+
 public enum State {
   ORBIT, ATTACK, SEARCH, DEFENCE, CONVERT, REFUEL; 
 }
