+++ src/main/java/swarm_wars_library/graphics/Render.java
@@ -1,8 +1,7 @@
-package swarm_wars_library.engine;
+package swarm_wars_library.graphics;
 
 import swarm_wars_library.engine.Vector2D;
 import swarm_wars_library.engine.Tag;
-import swarm_wars_library.engine.Particle;
 import java.util.*;
 
 import processing.core.PApplet;
