+++ src/test/java/swarm_wars_logic/comms/CommsTests.java
@@ -6,9 +6,9 @@ import java.beans.Transient;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import swarm_wars_library.engine.CommsGlobal;
-import swarm_wars_library.engine.CommsChannel;
-import swarm_wars_library.engine.CommsPacket;
+import swarm_wars_library.comms.CommsGlobal;
+import swarm_wars_library.comms.CommsChannel;
+import swarm_wars_library.comms.CommsPacket;
 import swarm_wars_library.engine.Vector2D;
 
 class CommsTests {
