--- src/test/java/swarm_wars_logic/engine/Template.java
@@ -1,18 +1,0 @@
-package swarm_wars_logic.engine;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.beans.Transient;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import swarm_wars_library.engine.RigidBody;
-import swarm_wars_library.engine.Vector2D;
-
-class TemplateTests {
-
-  @Test
-  @DisplayName("XXXX")
-  void XXXXXTest(){
-  }
-}
